+++ product_code/admission_service/impl/src/main/scala/de/upb/cs/uc4/admission/impl/AdmissionLoader.scala
@@ -1,0 +1,19 @@
+package de.upb.cs.uc4.admission.impl
+
+import com.lightbend.lagom.scaladsl.akka.discovery.AkkaDiscoveryComponents
+import com.lightbend.lagom.scaladsl.api.Descriptor
+import com.lightbend.lagom.scaladsl.devmode.LagomDevModeComponents
+import com.lightbend.lagom.scaladsl.server._
+import de.upb.cs.uc4.admission.api.AdmissionService
+
+class AdmissionLoader extends LagomApplicationLoader {
+
+  override def load(context: LagomApplicationContext): LagomApplication =
+    new AdmissionApplication(context) with AkkaDiscoveryComponents
+
+  override def loadDevMode(context: LagomApplicationContext): LagomApplication =
+    new AdmissionApplication(context) with LagomDevModeComponents
+
+  override def describeService: Some[Descriptor] = Some(readDescriptor[AdmissionService])
+}
+